♻️ refactor : 리뷰 점수 float 으로 변경

+++ src/main/java/com/seeat/server/domain/review/domain/entity/Review.java
@@ -36,12 +36,12 @@ public class Review extends BaseEntity {
 
     private String thumbnailUrl;
 
-    private double rating;
+    private float rating;
 
     private String content;
 
     /// 정적 팩토리 메서드
-    public static Review of(User user, Seat seat, String movieTitle, double rating, String content) {
+    public static Review of(User user, Seat seat, String movieTitle, float rating, String content) {
         return Review.builder()
                 .user(user)
                 .seat(seat)
